prove everything (*) without using "Total" on relations
I think that my proof on paper that the power objects are total is
incorrect; I spelunked it with Danny and it is almost certainly false in
our model.

On the bright side, using Jon Sterling Thought, I found a way to prove
everything in Coq without using totality except on natural numbers, and
it is true there.

I can explain more later.

(*) I proved every rule except that the later modality perserves sigma
types; we didn't need it in our examples. This rule should also be true
using the same method as I came up with; but I'll deal with it later.